pyproject.toml
[project]
name = "genai-for-engineers-class"
version = "0.2.0"
description = "Code examples for the GenAI for Engineers class"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic-haystack>=3.1.0",
    "autogen>=0.9.7",
    "diffusers>=0.35.2",
    "google-genai>=1.47.0",
    "guardrails-ai>=0.5.15",
    "haystack-ai>=2.15.2",
    "langfuse>=2.0.1, <3.0.0",
    "langfuse-haystack>=2.0.1",
    "litellm>=1.74.8",
    "numpy>=2.3.4",
    "openai>=1.97.1",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "polars>=1.35.1",
    "qdrant-client>=1.15.1",
    "qdrant-haystack>=9.2.0",
    "requests>=2.32.4",
    "scipy>=1.16.0",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
]

[dependency-groups]
ide = [
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.10",
]

[tool.uv]
package = true
